[HttpFoundation] use InputBag for Request::$request only if data is coming from a form
